% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_column_with_delimiter_separated_values.R
\name{tidy_column_with_delimiter_separated_values}
\alias{tidy_column_with_delimiter_separated_values}
\title{Tidy Column with Delimiter-Separated Values}
\usage{
tidy_column_with_delimiter_separated_values(data, delimiter, column)
}
\arguments{
\item{data}{A tibble/data frame containing the column to be tidied.}

\item{delimiter}{The delimiter used to separate values within the column.}

\item{column}{The name of the column that contains delimiter-separated values.}
}
\value{
A modified tibble/data frame where the specified column has been separated into multiple rows and pivoted to a wider format, showing the presence of each 'value.'
}
\description{
This function separates a column with delimiter-separated values into multiple rows and pivots it to wide format, indicating the presence of each value.
}
\examples{
example <- data.frame(ID = 1:3, Codes = c("A,B,C", "D,E", "F,G,H,I"))
tidy_column_with_delimiter_separated_values(example, delimiter = ",", column = Codes)
}
